Use AES256-CBC instead of RC4 by default
We used the PSK-RC4-SHA cipher suite by default, since the CBC ciphers
are broken in OpenSSL 1.0.1d.  However, many other applications don't
work (properly) when linked against 1.0.1d either, and it was quickly
replaced with 1.0.1e (which fixes the issue), so the number of 1.0.1d
users is probably very small.  Therefore, it should be safe to use the
PSK-AES256-CBC-SHA cipher by default.