etc/nsca-ng.cfg: Add a default "authorize" section
Add an "authorize" section which lets the server start out of the box
(instead of bailing out because the user didn't specify an "authorize"
section).  The new configuration still doesn't permit clients to submit
anything to Nagios (it just allows authenticated clients to talk to the
server), so the client will spit out a "not authorized" message until
the user edits the nsca-ng.cfg(5) file.