Fix test cases for GNU Hurd
GNU Hurd doesn't like the the trick we're using in tests/test_nsca.c to
avoid having to wait until the server notices that we opened the command
file for reading.  This means that each check might block for a few
seconds on Hurd, but that's still preferable to failing test cases.