Support specifying defaults for authorizations
Allow for specifying authorization settings outside of "authorize"
sections.  They will then serve as a fallback for "authorize" sections
that don't define these settings.

We don't use a libConfuse validation callback to implement this feature,
as it would then only work for default settings specified prior to the
corresponding "authorize" section(s).  This behaviour would be
especially problematic if multiple configuration files in a directory
tree are included recursively, as the order of inclusion is undefined.