Ignore errors when closing file descriptors
Don't check the return value of closefrom(3) or fcntl(3) when closing
open file descriptors on startup.  If there are no file descriptors to
close, closefrom(3) will fail with EBADF on NetBSD and OpenBSD.  On
FreeBSD and Solaris, closefrom(3) is actually declared to return void.

Thanks to Stuart Henderson for reporting this bug.